+++ app/src/main/java/za/kilowatch/ultimatefilemanager/util/MimeTypeHelper.kt
@@ -28,6 +28,7 @@ object MimeTypeHelper {
         "jxl"           -> "image/jxl"   // future-proofing
         "yaml", "yml"   -> "text/yaml"
         "m3u", "m3u8"   -> "audio/x-mpegurl"
+        "epub"          -> "application/epub+zip"
         else            -> "application/octet-stream"
     }
 }